ziti: Avoid shipping firmware
Starting ~ CPH2569_15.0.0.1301(EX01) firmware, any downgrades to
previous firmware bricks the device and puts the device in QDL mode.
Thus, its upto the user to be on any of the CPH2569_15.0.0.* firmwares

As a cautionary measure, any and all old builds are purged so that nobody
downloads and installs them.